# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "ProductController_getAllProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of products fetched successfully"
          }
        },
        "summary": "Get all products",
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        },
        "summary": "Create a new product",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductController_updateProduct",
        "parameters": [
          {
            "name": "product_id",
            "required": false,
            "in": "query",
            "description": "Product ID to update",
            "schema": {
              "example": "PROD123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          }
        },
        "summary": "Update product details",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/sellers": {
      "get": {
        "operationId": "SellerController_getDetails",
        "parameters": [],
        "responses": {
          "200": {
            "description": "seller details fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user profile details",
        "tags": [
          "Sellers"
        ]
      }
    },
    "/sellers/products": {
      "get": {
        "operationId": "SellerController_getAllProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Seller products fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all products of a seller",
        "tags": [
          "Sellers"
        ]
      }
    },
    "/sellers/{sid}/products": {
      "post": {
        "operationId": "SellerController_createProduct",
        "parameters": [
          {
            "name": "sid",
            "required": true,
            "in": "path",
            "description": "Seller ID",
            "schema": {
              "example": "94144541854e298c2d976cb893b81343",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add product for seller",
        "tags": [
          "Sellers"
        ]
      }
    },
    "/sellers/{sid}/products/delete": {
      "patch": {
        "operationId": "SellerController_deleteProduct",
        "parameters": [
          {
            "name": "sid",
            "required": true,
            "in": "path",
            "description": "seller ID",
            "schema": {
              "example": "acef497c1130f71ccfe63aaa1c9a607d",
              "type": "string"
            }
          },
          {
            "name": "pid",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "example": "beleza_saude",
              "type": "string"
            }
          },
          {
            "name": "product_name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete seller product",
        "tags": [
          "Sellers"
        ]
      }
    },
    "/sellers/{sid}/products/{pid}": {
      "patch": {
        "operationId": "SellerController_updateProduct",
        "parameters": [
          {
            "name": "sid",
            "required": true,
            "in": "path",
            "description": "seller ID",
            "schema": {
              "example": "05a360982be454c7e715c0fec4f67243",
              "type": "string"
            }
          },
          {
            "name": "pid",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "example": "05a360982be454c7e715c0fec4f67243",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update seller product",
        "tags": [
          "Sellers"
        ]
      }
    },
    "/sellers/{sid}/prodcts/{name}": {
      "get": {
        "operationId": "SellerController_getSearch",
        "parameters": [
          {
            "name": "sid",
            "required": true,
            "in": "path",
            "description": "Seller ID",
            "schema": {
              "example": "94144541854e298c2d976cb893b81343",
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Product name",
            "schema": {
              "example": "iphone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search products by name",
        "tags": [
          "Sellers"
        ]
      }
    },
...
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully"
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged out successfully"
          }
        },
        "summary": "Logout user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset link sent"
          }
        },
        "summary": "Send forgot password email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        },
        "summary": "Reset user password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/users/{uid}/products": {
      "get": {
        "operationId": "UserController_getAllProduct",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "Customer ID",
            "schema": {
              "example": "d3e7d37c0df9aef383f3f2a15b0dddfb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer products fetched successfully"
          }
        },
        "summary": "Get all products of a customer",
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UserController_getDetails",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "4f21938f7b925dd621343fc205395145"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details fetched successfully"
          }
        },
        "summary": "Get user profile details",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ecommerce API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "example": "PROD123"
          },
          "product_category_name": {
            "type": "string",
            "example": "electronics"
          },
          "product_photos_qty": {
            "type": "number",
            "example": 5
          },
          "product_weight_g": {
            "type": "number",
            "example": 500
          },
          "product_length_cm": {
            "type": "number",
            "example": 20
          },
          "product_height_cm": {
            "type": "number",
            "example": 10
          },
          "product_width_cm": {
            "type": "number",
            "example": 15
          }
        },
        "required": [
          "product_id",
          "product_category_name",
          "product_photos_qty",
          "product_weight_g",
          "product_length_cm",
          "product_height_cm",
          "product_width_cm"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "example": "PROD123"
          },
          "product_category_name": {
            "type": "string",
            "example": "electronics"
          },
          "product_photos_qty": {
            "type": "number",
            "example": 5
          },
          "product_weight_g": {
            "type": "number",
            "example": 500
          },
          "product_length_cm": {
            "type": "number",
            "example": 20
          },
          "product_height_cm": {
            "type": "number",
            "example": 10
          },
          "product_width_cm": {
            "type": "number",
            "example": 15
          }
        }
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "lipigava.22it@kongu.edu"
          },
          "username": {
            "type": "string",
            "example": "lipiga"
          },
          "password": {
            "type": "string",
            "example": "lipiga"
          },
          "zip_code": {
            "type": "number",
            "example": 600001
          },
          "city": {
            "type": "string",
            "example": "chennai"
          },
          "state": {
            "type": "string",
            "example": "TN"
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "zip_code",
          "city",
          "state"
        ]
      },
      "VerifyUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "lipigava.22it@kongu.edu"
          },
          "password": {
            "type": "string",
            "example": "lipiga"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "65f2a1b2c3d4"
          }
        },
        "required": [
          "userId"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "lipigava.22it@Kongu.edu"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyb2xlIjoidXNlciIsImVtYWlsIjoibGlwaWdhdmEuMjJpdEBrb25ndS5lZHUiLCJpYXQiOjE3NzI1MDY4OTYsImV4cCI6MTc3MzExMTY5Nn0.A528BU5WSx38KFU3Avs7HCKm95-BkPYVrNWZy_f4_wA"
          }
        },
        "required": [
          "email",
          "refreshToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "token": {
            "type": "string",
            "example": "reset-token-123"
          },
          "newPassword": {
            "type": "string",
            "example": "NewPass@123",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "token",
          "newPassword"
        ]
      },
...
    }
  }
}